requirements.txt
requests==2.31.0
python-dotenv==1.0.0
logtail-python==0.2.8

